initial GUI completion

+++ src/test/model/BeingTest.java
@@ -20,17 +20,17 @@ class BeingTest {
     @Test
     void collisionTest() {
         //new being that is at a different player, but same y pos. other cases are tested in subclasses
-        Being difPosPlayer1 = new Player(testPlayer.getPosX()+100, testPlayer.getPosY());
+        Being difPosPlayer1 = new Player((int) testPlayer.getPosX() + 100, (int) testPlayer.getPosY());
         assertFalse(testPlayer.collidedWith(difPosPlayer1));
         Being samePosPlayer = new Player(); // same position as player, in middle of gamestate
         assertTrue(testPlayer.collidedWith(samePosPlayer));
 
         // new being that has same x value but not y value
-        Being difPosPlayer2 = new Player(testPlayer.getPosX(), testPlayer.getPosY()+100);
+        Being difPosPlayer2 = new Player( (int) testPlayer.getPosX(), (int) testPlayer.getPosY()+100);
         assertFalse(testPlayer.collidedWith(difPosPlayer2));
 
         // new being that has no similar values
-        Being difPosPlayer3 = new Player(testPlayer.getPosX()+50, testPlayer.getPosY()+100);
+        Being difPosPlayer3 = new Player((int) testPlayer.getPosX() + 50, (int) testPlayer.getPosY() + 100);
         assertFalse(testPlayer.collidedWith(difPosPlayer3));
 
     }
